build: Improve portability for non-Linux systems.
This addresses compilation and build issues with
non-Linux systems, FreeBSD in particular.

The core BBS now builds successfully on FreeBSD
for me; some functionality is missing (namely
UUID functionality in utils.c and much functionality
in system.c). Most modules should compile as well,
though dependencies are not all addressed.

+++ nets/net_ssh.c
@@ -543,7 +543,7 @@ static int shell_request(ssh_session session, ssh_channel channel, void *userdat
 		return SSH_ERROR;
 	}
 	cdata->nodethread = node->thread;
-	bbs_debug(3, "Node thread is %lu\n", cdata->nodethread);
+	bbs_debug(3, "Node thread is %lu\n", (unsigned long) cdata->nodethread);
 	return SSH_OK;
 }
 
@@ -601,7 +601,7 @@ static inline int thread_has_exited(pthread_t thread)
 		return 1;
 	}
 	/* Unexpected return value */
-	bbs_warning("pthread_kill(%lu) = %d (%s)\n", thread, res, strerror(res));
+	bbs_warning("pthread_kill(%lu) = %d (%s)\n", (unsigned long) thread, res, strerror(res));
 	return 0;
 }
